requirements.txt
digit-interface
pillow
pyserial
